Add files via upload
Created React app via npx create-react-app MAST

For use, need to install node_modules and yarn

Currently have program directed to GPD Homepage on start, with a rough design. Have banner with Logout, and the Import dropdown menu, and the Add/Delete Student buttons that we need to add functionality to.